refactor: replace modal.js nad slide.js into shared folder (Thu, Dec 19, 2024 18:11)
