+++ src/assets/scripts/shared/slide.js
